Brian: Finishing pattern matching test example.

+++ languages/scala/src/main/scala/fp/java/InventoryItem.scala
@@ -1,3 +1,3 @@
 package fp.java
 
-case class InventoryItem(sku: String, desc: String, price: Float)
+case class InventoryItem(sku: String, desc: String, price: Double)
